requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.1.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
loguru==0.7.3
markupsafe==3.0.2
multidict==6.1.0
priority==2.0.0
propcache==0.3.0
python-dotenv==1.0.1
quart==0.20.0
quart-cors==0.8.0
quart-rate-limiter==0.11.0
ruff==0.9.10
tomli==2.2.1
werkzeug==3.1.3
win32-setctime==1.2.0
wsproto==1.2.0
yarl==1.18.3
